requirements.txt
aiogram==3.13.1
APScheduler==3.10.4
python-dotenv==1.0.1
ujson==5.10.0
